fix(engine): migrate matrix storage from float to double and strengthen freeform parsing
- Promote feature_matrix.values from float* to double* to prevent
  precision collapse on very small or very large numeric fields.
  All distance functions (euclidean_sq, manhattan, zscore_sq) and
  every clustering algorithm (kmeans++, dbscan, birch, mean_shift,
  gmm, agglomerative) are updated to operate on double vectors.

- When configured numeric_keys exist but fail to match a row
  entirely, rewind captured dimensions to 0 and re-init valid_mask
  so freeform number extraction can populate the matrix instead of
  leaving a row full of NaN placeholders. This fixes the entropy=0
  and trend=0 collapse seen when the config key list is stale.

- Add logana_strip_numeric_suffixes() that strips 40+ common unit
  suffixes (ms, %, MB, GB, Hz, bps, px, dpi, tps, qps, rpm, etc.)
  before strtod so unstructured text like latency=20.0ms or
  cpu=40% is accepted.

- Add logana_simd_sum_f64 companion to simd.h for future double
  SIMD reductions.

+++ include/logana/types.h
@@ -92,7 +92,7 @@ typedef struct {
 } logana_analysis_summary_t;
 
 typedef struct {
-    float *values;
+    double *values;
     uint64_t *timestamps;        /* Normalized epoch milliseconds per row */
     uint8_t *valid_mask;         /* Per-dimension validity: 1 = valid, 0 = invalid/parsed trap */
     uint8_t *formats;            /* Per-row format tag: 0=JSON, 1=KV, 2=Text */
